Switch lexer char iterator to CharIndices<>
This is because the [x..y] for rust strings indexes bytes, and i was
trying to index by char index which was causing some weird issues
when introducing larger-width characters in the file. now, the lexer
uses the index given by CharIndices iterator to index into
the file and for spans.